fix: Reubicado el botón de cambio de tema.
